ch99/mysite/requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
django-classy-tags==1.0.0
django-taggit==1.3.0
django-taggit-templatetags2==1.6.1
django-widget-tweaks==1.4.8
idna==2.9
lxml==4.5.1
pytz==2020.1
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
